Tarot refactored into Cog. Extracted API calls & message formatting to own files.
