--- src/tarot/magicEight.py
@@ -1,29 +1,0 @@
-import random
-
-
-async def magicEightBall(ctx, message=True):
-    if message:
-        eightBall = random.randint(0, 19)
-        outlooks = [
-            "As I see it, yes.",
-            "Ask again later.",
-            "Better not tell you now.",
-            "Cannot predict now.",
-            "Concentrate and ask again.",
-            "Don’t count on it.",
-            "It is certain.",
-            "It is decidedly so.",
-            "Most likely.",
-            "My reply is no.",
-            "My sources say no.",
-            "Outlook not so good.",
-            "Outlook good.",
-            "Reply hazy, try again.",
-            "Signs point to yes.",
-            "Very doubtful.",
-            "Without a doubt.",
-            "Yes.",
-            "Yes – definitely.",
-            "You may rely on it.",
-        ]
-        await ctx.send('Magic 8: ' + outlooks[eightBall])
